Bug fix around Int,Float types

+++ ddSubsetErrorModel/lnProbs.jl
@@ -5,10 +5,10 @@ module lnProbs
     using Distributions
     # x ∈ {0,1}, x = 1 w.p. p
     function ln_P_ber(x::INT, p::REAL)
-        return x * log(e, p) + (1-x) * log(e, (1-p))
+        return (REAL)(x * log(e, p) + (1-x) * log(e, (1-p)))
     end
 
     function ln_P_beta(p::REAL, α::REAL, β::REAL)
-        return (α-1)*log(e, p) + (β-1)*log(e,(1-p)) - lbeta(α, β)
+        return (REAL)((α-1)*log(e, p) + (β-1)*log(e,(1-p)) - lbeta(α, β))
     end
 end
